Moved course.java into model package
